Move elasticsearch publishing plugin to build-conventions
