+++ buildSrc/src/main/java/org/elasticsearch/gradle/internal/InternalPluginBuildPlugin.java
@@ -11,7 +11,7 @@ package org.elasticsearch.gradle.internal;
 import groovy.lang.Closure;
 import org.elasticsearch.gradle.internal.precommit.TestingConventionsTasks;
 import org.elasticsearch.gradle.internal.test.RestTestBasePlugin;
-import org.elasticsearch.gradle.internal.util.Util;
+import org.elasticsearch.gradle.internal.conventions.util.Util;
 import org.elasticsearch.gradle.plugin.PluginBuildPlugin;
 import org.elasticsearch.gradle.plugin.PluginPropertiesExtension;
 import org.elasticsearch.gradle.util.GradleUtils;
