+++ buildSrc/src/main/java/org/elasticsearch/gradle/internal/info/GlobalBuildInfoPlugin.java
@@ -10,7 +10,8 @@ package org.elasticsearch.gradle.internal.info;
 import org.apache.commons.io.IOUtils;
 import org.elasticsearch.gradle.internal.BwcVersions;
 import org.elasticsearch.gradle.OS;
-import org.elasticsearch.gradle.internal.util.Util;
+import org.elasticsearch.gradle.internal.conventions.info.GitInfo;
+import org.elasticsearch.gradle.internal.conventions.util.Util;
 import org.gradle.api.GradleException;
 import org.gradle.api.JavaVersion;
 import org.gradle.api.Plugin;
@@ -35,22 +36,17 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileReader;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.io.UncheckedIOException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Random;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -79,18 +75,18 @@ public class GlobalBuildInfoPlugin implements Plugin<Project> {
         if (project != project.getRootProject()) {
             throw new IllegalStateException(this.getClass().getName() + " can only be applied to the root project.");
         }
-        GradleVersion minimumGradleVersion = GradleVersion.version(Util.getResourceContents("/minimumGradleVersion"));
+        GradleVersion minimumGradleVersion = GradleVersion.version(getResourceContents("/minimumGradleVersion"));
         if (GradleVersion.current().compareTo(minimumGradleVersion) < 0) {
             throw new GradleException("Gradle " + minimumGradleVersion.getVersion() + "+ is required");
         }
 
-        JavaVersion minimumCompilerVersion = JavaVersion.toVersion(Util.getResourceContents("/minimumCompilerVersion"));
-        JavaVersion minimumRuntimeVersion = JavaVersion.toVersion(Util.getResourceContents("/minimumRuntimeVersion"));
+        JavaVersion minimumCompilerVersion = JavaVersion.toVersion(getResourceContents("/minimumCompilerVersion"));
+        JavaVersion minimumRuntimeVersion = JavaVersion.toVersion(getResourceContents("/minimumRuntimeVersion"));
 
         File runtimeJavaHome = findRuntimeJavaHome();
 
         File rootDir = project.getRootDir();
-        GitInfo gitInfo = gitInfo(rootDir);
+        GitInfo gitInfo = GitInfo.gitInfo(rootDir);
 
         BuildParams.init(params -> {
             // Initialize global build parameters
@@ -362,141 +358,24 @@ public class GlobalBuildInfoPlugin implements Plugin<Project> {
         return _defaultParallel;
     }
 
-    public static GitInfo gitInfo(File rootDir) {
-        try {
-            /*
-             * We want to avoid forking another process to run git rev-parse HEAD. Instead, we will read the refs manually. The
-             * documentation for this follows from https://git-scm.com/docs/gitrepository-layout and https://git-scm.com/docs/git-worktree.
-             *
-             * There are two cases to consider:
-             *  - a plain repository with .git directory at the root of the working tree
-             *  - a worktree with a plain text .git file at the root of the working tree
-             *
-             * In each case, our goal is to parse the HEAD file to get either a ref or a bare revision (in the case of being in detached
-             * HEAD state).
-             *
-             * In the case of a plain repository, we can read the HEAD file directly, resolved directly from the .git directory.
-             *
-             * In the case of a worktree, we read the gitdir from the plain text .git file. This resolves to a directory from which we read
-             * the HEAD file and resolve commondir to the plain git repository.
-             */
-            final Path dotGit = rootDir.toPath().resolve(".git");
-            final String revision;
-            if (Files.exists(dotGit) == false) {
-                return new GitInfo("unknown", "unknown");
-            }
-            final Path head;
-            final Path gitDir;
-            if (Files.isDirectory(dotGit)) {
-                // this is a git repository, we can read HEAD directly
-                head = dotGit.resolve("HEAD");
-                gitDir = dotGit;
-            } else {
-                // this is a git worktree, follow the pointer to the repository
-                final Path workTree = Paths.get(readFirstLine(dotGit).substring("gitdir:".length()).trim());
-                if (Files.exists(workTree) == false) {
-                    return new GitInfo("unknown", "unknown");
+
+    public static String getResourceContents(String resourcePath) {
+        try (
+                BufferedReader reader = new BufferedReader(new InputStreamReader(GlobalBuildInfoPlugin.class.getResourceAsStream(resourcePath)))
+        ) {
+            StringBuilder b = new StringBuilder();
+            for (String line = reader.readLine(); line != null; line = reader.readLine()) {
+                if (b.length() != 0) {
+                    b.append('\n');
                 }
-                head = workTree.resolve("HEAD");
-                final Path commonDir = Paths.get(readFirstLine(workTree.resolve("commondir")));
-                if (commonDir.isAbsolute()) {
-                    gitDir = commonDir;
-                } else {
-                    // this is the common case
-                    gitDir = workTree.resolve(commonDir);
-                }
+                b.append(line);
             }
-            final String ref = readFirstLine(head);
-            if (ref.startsWith("ref:")) {
-                String refName = ref.substring("ref:".length()).trim();
-                Path refFile = gitDir.resolve(refName);
-                if (Files.exists(refFile)) {
-                    revision = readFirstLine(refFile);
-                } else if (Files.exists(gitDir.resolve("packed-refs"))) {
-                    // Check packed references for commit ID
-                    Pattern p = Pattern.compile("^([a-f0-9]{40}) " + refName + "$");
-                    try (Stream<String> lines = Files.lines(gitDir.resolve("packed-refs"))) {
-                        revision = lines.map(p::matcher)
-                            .filter(Matcher::matches)
-                            .map(m -> m.group(1))
-                            .findFirst()
-                            .orElseThrow(() -> new IOException("Packed reference not found for refName " + refName));
-                    }
-                } else {
-                    File refsDir = gitDir.resolve("refs").toFile();
-                    if (refsDir.exists()) {
-                        String foundRefs = Arrays.stream(refsDir.listFiles()).map(f -> f.getName()).collect(Collectors.joining("\n"));
-                        Logging.getLogger(GlobalBuildInfoPlugin.class).error("Found git refs\n" + foundRefs);
-                    } else {
-                        Logging.getLogger(GlobalBuildInfoPlugin.class).error("No git refs dir found");
-                    }
-                    throw new GradleException("Can't find revision for refName " + refName);
-                }
-            } else {
-                // we are in detached HEAD state
-                revision = ref;
-            }
-            return new GitInfo(revision, findOriginUrl(gitDir.resolve("config")));
-        } catch (final IOException e) {
-            // for now, do not be lenient until we have better understanding of real-world scenarios where this happens
-            throw new GradleException("unable to read the git revision", e);
+
+            return b.toString();
+        } catch (IOException e) {
+            throw new UncheckedIOException("Error trying to read classpath resource: " + resourcePath, e);
         }
     }
 
-    private static String findOriginUrl(final Path configFile) throws IOException {
-        Map<String, String> props = new HashMap<>();
 
-        try (Stream<String> stream = Files.lines(configFile, StandardCharsets.UTF_8)) {
-            Iterator<String> lines = stream.iterator();
-            boolean foundOrigin = false;
-            while (lines.hasNext()) {
-                String line = lines.next().trim();
-                if (line.startsWith(";") || line.startsWith("#")) {
-                    // ignore comments
-                    continue;
-                }
-                if (foundOrigin) {
-                    if (line.startsWith("[")) {
-                        // we're on to the next config item so stop looking
-                        break;
-                    }
-                    String[] pair = line.trim().split("=", 2);
-                    props.put(pair[0].trim(), pair[1].trim());
-                } else {
-                    if (line.equals("[remote \"origin\"]")) {
-                        foundOrigin = true;
-                    }
-                }
-            }
-        }
-
-        String originUrl = props.get("url");
-        return originUrl == null ? "unknown" : originUrl;
-    }
-
-    private static String readFirstLine(final Path path) throws IOException {
-        String firstLine;
-        try (Stream<String> lines = Files.lines(path, StandardCharsets.UTF_8)) {
-            firstLine = lines.findFirst().orElseThrow(() -> new IOException("file [" + path + "] is empty"));
-        }
-        return firstLine;
-    }
-
-    public static class GitInfo {
-        private final String revision;
-        private final String origin;
-
-        GitInfo(String revision, String origin) {
-            this.revision = revision;
-            this.origin = origin;
-        }
-
-        public String getRevision() {
-            return revision;
-        }
-
-        public String getOrigin() {
-            return origin;
-        }
-    }
 }
